\name{kFNews}
\alias{kFNews}
\title{Display the NEWS file}
\usage{
kFNews()
}
\value{
None.
}
\description{
\code{kFNews} shows the NEWS file of the kernelFactory
package.
}
\examples{
kFNews()
}
\author{
Authors: Michel Ballings and Dirk Van den Poel, Maintainer:
\email{Michel.Ballings@GMail.com}
}
\references{
Ballings, M. and Van den Poel, D. (2013), Kernel Factory:
An Ensemble of Kernel Machines. Expert Systems With
Applications, 40(8), 2904-2913.
}
\seealso{
\code{\link{kernelFactory}},
\code{\link{predict.kernelFactory}}
}

